requirements.txt
fastapi
uvicorn
llama-cpp-python
faiss-cpu
sentence-transformers
torch
python-dotenv
google-generativeai
